+++ lib/clixon/clixon_netconf_lib.h
@@ -244,5 +244,6 @@ int netconf_output(int s, cbuf *xf, const char *msg);
 int netconf_output_encap(netconf_framing_type framing, cbuf *cb);
 int netconf_input_chunked_framing(char ch, int *state, size_t *size);
 int netconf_cbuf_err2cb(clixon_handle h, cbuf *cb0, cbuf *cberr);
+char *netconf_reply_err_tag(cxobj *xret);
 
 #endif /* _CLIXON_NETCONF_LIB_H */